Made changes related to the pr reviews
