+++ gradle.properties
@@ -1,13 +1,14 @@
 # Check these on https://modmuss50.me/fabric.html
-kotlin.code.style=official
-org.gradle.jvmargs=-Xmx2G
-minecraft_version=1.16.3
-yarn_mappings=1.16.3+build.1
-loader_version=0.9.3+build.207
-fabric_version=0.20.2+build.402-1.16
-loom_version=0.2.7-SNAPSHOT
+kotlin.code.style = official
+org.gradle.jvmargs = -Xmx2G
+minecraft_version = 1.16.3
+yarn_mappings = 1.16.3+build.1
+loader_version = 0.9.3+build.207
+fabric_version = 0.20.2+build.402-1.16
+loom_version = 0.2.7-SNAPSHOT
 mod_version = 1.0.0
 maven_group = dev.toastmc
 archives_base_name = toastclient
-kotlin_version=1.4.10
+kotlin_version = 1.4.10
 fiber_version = 0.23.0
+multiconnect = 1.3.19
