+++ src/main/java/dev/toastmc/client/command/util/CommandManager.kt
@@ -13,7 +13,7 @@ class CommandManager {
         commands.clear()
         commands.addAll(listOf(
                 Help(), Hide(), Pos(),
-                Prefix(), Toggle()
+                Prefix(), Teleport(), Toggle()
         ))
         commands.forEach {
             it.register(Command.dispatcher)
