+++ src/main/java/dev/toastmc/client/util/Util.kt
@@ -716,6 +716,7 @@ object KeyUtil {
     }
 
     init {
+//        keys["!"] =
         keys[" "] = 32
         keys["'"] = 39
         keys[","] = 44
